% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/malikar.R
\docType{data}
\name{cars2}
\alias{cars2}
\title{Historical data on cars}
\format{A data frame with 50 rows and 2 variables:
\describe{
  \item{speed}{...}
  \item{dist}{0000}
  ...
}}
\source{
\url{www.yandex.ru}
}
\usage{
cars2
}
\description{
A dataset containing speed and distance
}
\keyword{datasets}

