% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/malikar.R
\name{malika}
\alias{malika}
\title{square number}
\usage{
malika(x, n = 2)
}
\arguments{
\item{x}{the number that will be squared}

\item{n}{the power (default is 2)}
}
\value{
x power n, x^n
}
\description{
function that squares numbers
}
\details{
This important function may be used to
square any number :)
}
\examples{
malika(5)
malika(5,n=6)
}

